fix(test): split sdk-adapter-dispositions to stay under CI file-timeout (#4475)
sdk-adapter-dispositions.test.ts exited 0 but exceeded the CI 300s
file-timeout at ~489s. Instrumented fresh-process measurement (bun's
own accounting: 488.80s in tests, ~200ms from last test to process
exit) proved all runtime is genuine per-fixture cost — 291 production
SDK host startups at ~1.7s each — not leaked timers/sockets/workers.

Split the three machine-adapter cohorts (mcp, acp, daemonCli) into
sibling files so each runs under the unchanged timeout with
byte-identical coverage. Chat-adapter loops (no production host, ~1s)
remain in the original file.

Measured runtimes per split file (preload + 60s per-test timeout):
  chat:   291 tests,  1.1s
  mcp:     97 tests, 164.5s
  acp:     97 tests, 156.1s
  daemonCli: 97 tests, 178.2s

Updated generate-sdk-adapter-parity-manifest.ts and
telegram-baseline-manifest.test.ts for the new file layout. No tests
skipped, no timeout inflated.

Lore-id: 4475-sdk-adapter-split
Constraint: must not globally inflate CI file-timeout or skip tests
Constraint: coverage must remain byte-identical (582 parity rows)
Rejected: beforeAll fixture sharing | changes test isolation semantics
Rejected: inflating file-timeout to 600s | masks genuine cost growth
Confidence: high
Scope-risk: narrow
Reversibility: merge-safe
Tested: all 4 split files pass under exact CI harness args
Not-tested: CI shard distribution (requires push)
Supersedes: none

Closes #4475
