+++ packages/coding-agent/CHANGELOG.md
@@ -13,6 +13,7 @@
 - `gjc_coordinator_stop_session` no longer reports `close_failed` after a successful DR-1 terminal close. Reap now proves the same retained session is `terminal` and non-`live` for the exact `workspace/generation/incarnation` before completing local cleanup; rotated generation, different incarnation, ambiguous, still-live, and `terminal_uncertain` remain fail-closed (#4431).
 - The crash signature index can no longer quarantine itself. `parseCrashIndex` rejects an entry whose retained count exceeds its lifetime count, and compaction produced exactly that whenever the crash log still held more identity-bearing records for a signature than the journal had counted — which happens whenever a fatal's journal append fails or spends the per-process latch while its crash-log record is still written. The next read then discarded the whole index, taking every signature in it. The retained count is now held at the journaled lifetime.
 - The mid-run context maintenance contract suite no longer uses real-time `Bun.sleep(10)` to settle persistence after externally emitted events. A new public `AgentSession.awaitSessionSettlement()` deterministically awaits every in-flight agent-event handler (including the synchronous canonical persistence each one performs) and the durable transcript flush queue, replacing the ambiguous fixed-duration sleep whose adequacy depended on system load. Concurrent-state observation helpers now use event-loop tick yields (`Bun.sleep(0)`) instead of paced real-time sleeps. The maintenance outcome contract (`not-needed | pruned | compacted | promoted | failed | aborted`), publication failure semantics (fail-closed `artifact_persistence`), rollback, and canonical-output retention are unchanged (#4476).
+- `sdk-adapter-dispositions.test.ts` no longer exceeds the CI 300s file-timeout with `exit=0`. Measurement proved all 489s was genuine per-fixture runtime (291 production SDK host startups at ~1.7s each), not leaked timers/sockets/workers — the process exits ~200ms after the last test. The three machine-adapter cohorts (MCP, ACP, daemon CLI) are now split into sibling files so each runs under the unchanged timeout with byte-identical coverage; the chat-adapter loops stay in the original file (#4475).
 - Print mode now requests a governed process exit after successful session teardown, so completed one-shot runs cannot be pinned by residual runtime handles; both stdout and stderr are drained first and the recorded exit code is preserved.
 - `todo_write` no longer rejects a positional task handle without saying how to address a task. Callers reach for `id`/`index` because the tool result renders todos as a list, and `todo-write.ts` already had a message for exactly that confusion, but for model-issued calls raw validation rejects the unknown key before `execute` runs (only bridges that invoke a tool directly, such as `eval`'s `callSessionTool`, still reach it), so the model only saw `rejected key: "id"` next to the legal-key list and had to guess the replacement (each retry discards the whole batch, since rejection is atomic). The positional-handle family (`id`, `ids`, `index`, `taskId`, `task_id`), including one nested under `task` or `content`, now carries the same correction the executor gives: tasks have no id or index, target one with `task` set to its exact content or a whole phase with `phase`. A correction shared by several rejected keys is stated once, and raw diagnostics retain only eight keys with a deterministic omitted-key count. The executor comment claiming those keys are stripped before it runs was stale and is corrected.
 - Post-merge repair for #4401/#4399: `CHAT_DAEMON_GENERATIONS.discord` 63→64 and `.slack` 66→67 so the shared `SessionRouter` attachment lifecycle changes from #4401 (`#attach`, `#createAttachedClient`, `#publishAttachment`) are properly generation-fenced for Discord and Slack daemons, not just Telegram. The semantic guard manifest is regenerated.
